Made the make posting more functional
The make posting is now much more functional than it was. Made new components to add other users to a group lesson, and to select multiple dates for the lesson. Issue now is when editing, the already selected dates are passed through as strings as opposed to a date object